hyperliquid-bot/requirements.txt
# Hyperliquid Trading Bot Dependencies
hyperliquid-python-sdk>=0.5.0
requests>=2.31.0
websockets>=12.0
pandas>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
